fix(video-pipeline): move TransmissionManager to transmission package, delete .gitkeep
