+++ src/pipeline/video/index.ts
@@ -5,6 +5,7 @@ export {
   type RawVideoFrame,
 } from './LandmarkExtractor';
 export { TrackingHealthMonitor } from './TrackingHealthMonitor';
-export { VideoPipeline, type IVideoPipeline, type RawMediaHandle, type TransmissionManager } from './VideoPipeline';
+export { VideoPipeline, type IVideoPipeline, type RawMediaHandle } from './VideoPipeline';
+export type { TransmissionManager } from '../../transmission';
 export type { LandmarkFrame } from '../../common/models';
