+++ src/transmission/index.ts
@@ -1,0 +1,1 @@
+export type { TransmissionManager } from './TransmissionManager';
